fix: Change PK datatype Integer to Long (#4)

+++ src/main/java/com/m1s/m1sserver/Repository/CounselSolutionRepository.java
@@ -3,5 +3,5 @@ package com.m1s.m1sserver.Repository;
 import com.m1s.m1sserver.Model.CounselSolution;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CounselSolutionRepository extends JpaRepository<CounselSolution, Integer> {
+public interface CounselSolutionRepository extends JpaRepository<CounselSolution, Long> {
 }
